nits fix
