+++ packages/nim-serving/src/constants.ts
@@ -1,0 +1,1 @@
+export const NIM_OPERATOR_MANAGED_BY = 'k8s-nim-operator';
