+++ packages/nim-serving/src/pages/deployments/NIMServingDetails.tsx
